subsys: shell: Rename shell to legacy_shell and cli to shell
Added legacy_ prefix to shell.h and shell.c and aligned zephyr to that
change. Renamed cli.[ch] to shell.[ch] and updated includes accordingly.
