+++ include/shell/shell_uart.h
@@ -7,7 +7,7 @@
 #ifndef SHELL_UART_H__
 #define SHELL_UART_H__
 
-#include <shell/cli.h>
+#include <shell/shell.h>
 
 #ifdef __cplusplus
 extern "C" {
