+++ subsys/bluetooth/shell/ll.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <zephyr.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <misc/printk.h>
 
 #include <bluetooth/hci.h>
