+++ subsys/net/lib/openthread/platform/shell.c
@@ -7,7 +7,7 @@
 #include <kernel.h>
 #include <stdio.h>
 #include <misc/printk.h>
-#include <shell/shell.h>
+#include <shell/legacy_shell.h>
 #include <openthread/cli.h>
 #include <platform.h>
