+++ subsys/shell/shell.c
@@ -1,652 +1,1861 @@
 /*
- * Copyright (c) 2015 Intel Corporation
+ * Copyright (c) 2018 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/**
- * @file
- * @brief Console handler implementation of shell.h API
- */
-
-#include <zephyr.h>
-#include <stdio.h>
-#include <string.h>
-#include <version.h>
-
-#include <console.h>
-#include <drivers/console/console.h>
-#include <misc/printk.h>
-#include <misc/util.h>
-#include "mgmt/serial.h"
-
+#include <ctype.h>
+#include <stdarg.h>
+#include <stdlib.h>
 #include <shell/shell.h>
+#include "shell_utils.h"
+#include "shell_ops.h"
+#include "shell_wildcard.h"
+#include "shell_vt100.h"
+#include <assert.h>
+#include <atomic.h>
+#include <logging/log_backend.h>
+#include <logging/log_ctrl.h>
+#include <logging/log_output.h>
+#include <lib/fnmatch/fnmatch.h>
+#include <logging/log.h>
 
-#if defined(CONFIG_NATIVE_POSIX_CONSOLE)
-#include "drivers/console/native_posix_console.h"
+#define LOG_MODULE_NAME shell
+#define LOG_LEVEL CONFIG_SHELL_LOG_LEVEL
+LOG_MODULE_REGISTER();
+
+/* 2 == 1 char for cmd + 1 char for '\0' */
+#if (CONFIG_SHELL_CMD_BUFF_SIZE < 2)
+	#error too small CONFIG_SHELL_CMD_BUFF_SIZE
 #endif
 
-#define ARGC_MAX 10
-#define COMMAND_MAX_LEN 50
-#define MODULE_NAME_MAX_LEN 20
-/* additional chars are " >" (include '\0' )*/
-#define PROMPT_SUFFIX 3
-#define PROMPT_MAX_LEN (MODULE_NAME_MAX_LEN + PROMPT_SUFFIX)
+#if (CONFIG_SHELL_PRINTF_BUFF_SIZE < 1)
+	#error too small SHELL_PRINTF_BUFF_SIZE
+#endif
 
-/* command table is located in the dedicated memory segment (.shell_) */
-extern struct shell_module __shell_module_start[];
-extern struct shell_module __shell_module_end[];
+#define SHELL_MSG_COMMAND_NOT_FOUND	": command not found"
+#define SHELL_MSG_TAB_OVERFLOWED	\
+	"Tab function: commands counter overflowed.\r\n"
 
-extern struct shell_cmd __shell_cmd_start[];
-extern struct shell_cmd __shell_cmd_end[];
+#define CLI_DATA_SECTION_ITEM_GET(i)	\
+	NRF_SECTION_ITEM_GET(cli_command, shell_cmd_entry_t, (i))
+#define CLI_DATA_SECTION_ITEM_COUNT	\
+	NRF_SECTION_ITEM_COUNT(cli_command, shell_cmd_entry_t)
 
-#define NUM_OF_SHELL_ENTITIES (__shell_module_end - __shell_module_start)
-#define NUM_OF_SHELL_CMDS (__shell_cmd_end - __shell_cmd_start)
+#define CLI_SORTED_CMD_PTRS_ITEM_GET(i)	\
+	NRF_SECTION_ITEM_GET(cli_sorted_cmd_ptrs, const char *, (i))
+#define CLI_SORTED_CMD_PTRS_START_ADDR_GET	\
+	NRF_SECTION_START_ADDR(cli_sorted_cmd_ptrs)
 
-static const char *prompt;
-static char default_module_prompt[PROMPT_MAX_LEN];
-static struct shell_module *default_module;
-static bool no_promt;
+#define SHELL_INIT_OPTION_PRINTER	(NULL)
 
-#define STACKSIZE CONFIG_CONSOLE_SHELL_STACKSIZE
-static K_THREAD_STACK_DEFINE(stack, STACKSIZE);
-static struct k_thread shell_thread;
+/* Initial cursor position is: (1, 1). */
+#define SHELL_INITIAL_CURS_POS		(1u)
 
-#define MAX_CMD_QUEUED CONFIG_CONSOLE_SHELL_MAX_CMD_QUEUED
-static struct console_input buf[MAX_CMD_QUEUED];
 
-static struct k_fifo avail_queue;
-static struct k_fifo cmds_queue;
+/*static bool shell_log_entry_process(const struct shell *shell, bool skip);*/
+static void shell_execute(const struct shell *shell);
 
-static shell_cmd_function_t app_cmd_handler;
-static shell_prompt_function_t app_prompt_handler;
+extern const struct shell_cmd_entry __shell_root_cmds_start[0];
+extern const struct shell_cmd_entry __shell_root_cmds_end[0];
 
-static shell_mcumgr_function_t mcumgr_cmd_handler;
-static void *mcumgr_arg;
-
-static const char *get_prompt(void)
+static inline const struct shell_cmd_entry *shell_root_cmd_get(u32_t id)
 {
-	if (app_prompt_handler) {
-		const char *str;
+	return &__shell_root_cmds_start[id];
+}
 
-		str = app_prompt_handler();
-		if (str) {
-			return str;
+static inline u32_t shell_root_cmd_count(void)
+{
+	return ((void *)__shell_root_cmds_end -
+			(void *)__shell_root_cmds_start)/
+				sizeof(struct shell_cmd_entry);
+}
+
+static inline void transport_buffer_flush(const struct shell *shell) {
+	shell_fprintf_buffer_flush(shell->fprintf_ctx);
+}
+
+static inline void flag_help_set(const struct shell *shell)
+{
+	shell->ctx->internal.flags.show_help = 1;
+}
+static inline void flag_help_clear(const struct shell *shell)
+{
+	shell->ctx->internal.flags.show_help = 0;
+}
+
+static inline bool flag_processing_is_set(const struct shell *shell)
+{
+	return shell->ctx->internal.flags.processing == 1 ? true : false;
+}
+
+static inline void receive_state_change(const struct shell *shell,
+					enum shell_receive_state state)
+{
+	shell->ctx->receive_state = state;
+}
+
+static void cli_cmd_buffer_clear(const struct shell *shell)
+{
+	shell->ctx->cmd_buff[0] = '\0'; /* clear command buffer */
+	shell->ctx->cmd_buff_pos = 0;
+	shell->ctx->cmd_buff_len = 0;
+}
+
+/* Function sends data stream to the CLI instance. Each time before the
+ * cli_write function is called, it must be ensured that IO buffer of fprintf is
+ * flushed to avoid synchronization issues.
+ * For that purpose, use function transport_buffer_flush(shell)
+ */
+static void shell_write(const struct shell *shell, const void *data,
+			size_t length)
+{
+	assert(shell && data);
+
+	size_t offset = 0;
+	size_t tmp_cnt;
+
+	while (length) {
+		int err = shell->iface->api->write(shell->iface,
+				&((u8_t const *) data)[offset], length,
+				&tmp_cnt);
+		(void)err;
+		assert(err == 0);
+		assert(length >= tmp_cnt);
+		offset += tmp_cnt;
+		length -= tmp_cnt;
+		if (tmp_cnt == 0 &&
+		    (shell->ctx->state != SHELL_STATE_PANIC_MODE_ACTIVE)) {
+			/* todo  semaphore pend*/
+			if (IS_ENABLED(CONFIG_MULTITHREADING)) {
+				k_poll(&shell->ctx->events[SHELL_SIGNAL_TXDONE],
+				1, K_FOREVER);
+			} else {
+				/* Blocking wait in case of bare metal. */
+				while (!shell->ctx->internal.flags.tx_rdy) {
+
+				}
+				shell->ctx->internal.flags.tx_rdy = 0;
+			}
+		}
+	}
+}
+
+/* @brief Function shall be used to search commands.
+ *
+ * It moves the pointer entry to command of static command structure. If the
+ * command cannot be found, the function will set entry to NULL.
+ *
+ *   @param command	Pointer to command which will be processed (no matter
+ *			the root command).
+ *   @param lvl		Level of the requested command.
+ *   @param idx		Index of the requested command.
+ *   @param entry	Pointer which points to subcommand[idx] after function
+ *			execution.
+ *   @param st_entry	Pointer to the structure where dynamic entry data can be
+ *			stored.
+ */
+static void cmd_get(const struct shell_cmd_entry *command, size_t lvl,
+		    size_t idx, const struct shell_static_entry **entry,
+		    struct shell_static_entry *d_entry) {
+	assert(entry != NULL);
+	assert(st_entry != NULL);
+
+	if (lvl == SHELL_CMD_ROOT_LVL) {
+		if (idx < shell_root_cmd_count()) {
+			const struct shell_cmd_entry *cmd;
+
+			cmd = shell_root_cmd_get(idx);
+			*entry = cmd->u.entry;
+		} else {
+			*entry = NULL;
+		}
+		return;
+	}
+
+	if (command == NULL) {
+		*entry = NULL;
+		return;
+	}
+
+	if (command->is_dynamic) {
+		command->u.dynamic_get(idx, d_entry);
+		*entry = (d_entry->syntax != NULL) ? d_entry : NULL;
+	} else {
+		*entry = (command->u.entry[idx].syntax != NULL) ?
+				&command->u.entry[idx] : NULL;
+	}
+}
+
+static void vt100_color_set(const struct shell *shell,
+			    enum shell_vt100_color color)
+{
+
+	if (shell->ctx->vt100_ctx.col.col == color)
+	{
+		return;
+	}
+
+	shell->ctx->vt100_ctx.col.col = color;
+
+	if (color != SHELL_DEFAULT) {
+
+		u8_t cmd[] = SHELL_VT100_COLOR(color - 1);
+
+		shell_raw_fprintf(shell->fprintf_ctx, "%s", cmd);
+	}
+	else {
+		static u8_t const cmd[] = SHELL_VT100_MODESOFF;
+
+		shell_raw_fprintf(shell->fprintf_ctx, "%s", cmd);
+	}
+}
+
+static void vt100_bgcolor_set(const struct shell *shell,
+			      enum shell_vt100_color bgcolor)
+{
+	if ((bgcolor == SHELL_DEFAULT) ||
+	    (shell->ctx->vt100_ctx.col.bgcol == bgcolor)) {
+		return;
+	}
+
+	/* -1 because default value is first in enum */
+	u8_t cmd[] = SHELL_VT100_BGCOLOR(bgcolor - 1);
+
+	shell->ctx->vt100_ctx.col.bgcol = bgcolor;
+	shell_raw_fprintf(shell->fprintf_ctx, "%s", cmd);
+
+}
+
+static inline void vt100_colors_store(const struct shell *shell,
+				      struct shell_vt100_colors *color)
+{
+	memcpy(color, &shell->ctx->vt100_ctx.col, sizeof(*color));
+}
+
+static void vt100_colors_restore(const struct shell *shell,
+				 struct shell_vt100_colors const *color)
+{
+	vt100_color_set(shell, color->col);
+	vt100_bgcolor_set(shell, color->bgcol);
+}
+
+static void shell_state_set(const struct shell *shell, enum shell_state state)
+{
+	shell->ctx->state = state;
+
+	if (state == SHELL_STATE_ACTIVE) {
+		cli_cmd_buffer_clear(shell);
+		shell_fprintf(shell, SHELL_INFO, "%s", shell->name);
+	}
+}
+
+static void option_print(const struct shell *shell, char const *option,
+			 u16_t longest_option) {
+	static char const *tab = "  ";
+	u16_t columns;
+	u16_t diff;
+
+	/* Function initialization has been requested. */
+	if (option == NULL) {
+		shell->ctx->vt100_ctx.printed_cmd = 0;
+		return;
+	}
+
+	longest_option += shell_strlen(tab);
+
+	columns = (shell->ctx->vt100_ctx.cons.terminal_wid
+			- shell_strlen(tab)) / longest_option;
+	diff = longest_option - shell_strlen(option);
+
+	if (shell->ctx->vt100_ctx.printed_cmd++ % columns == 0) {
+		shell_fprintf(shell, SHELL_OPTION, "\r\n%s%s", tab, option);
+	} else {
+		shell_fprintf(shell, SHELL_OPTION, "%s", option);
+	}
+
+	shell_op_cursor_horiz_move(shell, diff);
+}
+
+static void history_init(const struct shell *shell)
+{
+	if (!IS_ENABLED(CONFIG_SHELL_HISTORY)) {
+		return;
+	}
+
+	shell_history_init(shell->history);
+}
+
+static void history_purge(const struct shell *shell)
+{
+	if (!IS_ENABLED(CONFIG_SHELL_HISTORY)) {
+		return;
+	}
+
+	shell_history_purge(shell->history);
+}
+
+static void history_mode_exit(const struct shell *shell)
+{
+	if (!IS_ENABLED(CONFIG_SHELL_HISTORY)) {
+		return;
+	}
+
+	shell_history_mode_exit(shell->history);
+}
+
+static void history_put(const struct shell *shell, u8_t *line, size_t length)
+{
+	if (!IS_ENABLED(CONFIG_SHELL_HISTORY)) {
+		return;
+	}
+
+	shell_history_put(shell->history, line, length);
+}
+
+static void history_handle(const struct shell *shell, bool up)
+{
+	size_t len;
+	bool history_mode;
+
+	/*optional feature */
+	if (!IS_ENABLED(CONFIG_SHELL_HISTORY)) {
+		return;
+	}
+
+	/* Backup command if history is entered */
+	if(!shell_history_active(shell->history)) {
+		if (up) {
+			u16_t cmd_len = shell_strlen(shell->ctx->cmd_buff);
+
+			if (cmd_len) {
+				strcpy(shell->ctx->temp_buff,
+				       shell->ctx->cmd_buff);
+			} else {
+				shell->ctx->temp_buff[0] = '\0';
+			}
+		} else {
+			/* Pressing 'down' not in history mode has no effect. */
+			return;
 		}
 	}
 
-	if (default_module) {
-		if (default_module->prompt) {
-			const char *ret;
+	/* Start by checking if history is not empty. */
+	history_mode = shell_history_get(shell->history, true,
+					 shell->ctx->cmd_buff, &len);
 
-			ret = default_module->prompt();
-			if (ret) {
-				return ret;
+	/* On exiting history mode print backed up command. */
+	if (!history_mode) {
+		strcpy(shell->ctx->cmd_buff, shell->ctx->temp_buff);
+		len = shell_strlen(shell->ctx->cmd_buff);
+	}
+
+	if (len) {
+		shell_op_cursor_home_move(shell);
+		clear_eos(shell);
+		shell_fprintf(shell, SHELL_NORMAL, "%s", shell->ctx->cmd_buff);
+		shell->ctx->cmd_buff_pos = len;
+		shell->ctx->cmd_buff_len = len;
+		shell_op_cond_next_line(shell);
+	}
+}
+
+static const struct shell_static_entry *find_cmd(
+					     const struct shell_cmd_entry *cmd,
+					     size_t lvl,
+					     char *cmd_str,
+					     struct shell_static_entry *d_entry)
+{
+	const struct shell_static_entry *entry = NULL;
+	size_t idx = 0;
+
+	do {
+		cmd_get(cmd, lvl, idx++, &entry, d_entry);
+		if (entry && (strcmp(cmd_str, entry->syntax) == 0)) {
+			LOG_INF("match %s %s", cmd_str, entry->syntax);
+			return entry;
+		}
+	} while (entry);
+
+	return entry;
+}
+
+/** @brief Function for getting last valid command in list of arguments. */
+static const struct shell_static_entry *get_last_command(
+					     const struct shell *shell,
+					     size_t argc,
+					     char *argv[],
+					     size_t *match_arg,
+					     struct shell_static_entry *d_entry)
+{
+	const struct shell_cmd_entry *prev_cmd = NULL;
+	const struct shell_static_entry *entry = NULL;
+	const struct shell_static_entry *prev_entry = NULL;
+	*match_arg = SHELL_CMD_ROOT_LVL;
+
+	while (*match_arg < argc) {
+
+		if (IS_ENABLED(CONFIG_SHELL_WILDCARD)) {
+			/* ignore wildcard argument */
+			if (shell_wildcard_character_exist(argv[*match_arg])) {
+				(*match_arg)++;
+				continue;
 			}
 		}
 
-		return default_module_prompt;
-	}
-
-	return prompt;
-}
-
-static void line_queue_init(void)
-{
-	int i;
-
-	for (i = 0; i < MAX_CMD_QUEUED; i++) {
-		k_fifo_put(&avail_queue, &buf[i]);
-	}
-}
-
-static size_t line2argv(char *str, char *argv[], size_t size)
-{
-	size_t argc = 0;
-
-	if (!strlen(str)) {
-		return 0;
-	}
-
-	while (*str && *str == ' ') {
-		str++;
-	}
-
-	if (!*str) {
-		return 0;
-	}
-
-	argv[argc++] = str;
-
-	while ((str = strchr(str, ' '))) {
-		*str++ = '\0';
-
-		while (*str && *str == ' ') {
-			str++;
+		entry = find_cmd(prev_cmd, *match_arg, argv[*match_arg],
+				 d_entry);
+		if (entry) {
+			prev_cmd = entry->subcmd;
+			prev_entry = entry;
+			(*match_arg)++;
+		} else {
+			entry = NULL;
+			break;
 		}
+	}
 
-		if (!*str) {
+	return entry;
+}
+
+static inline u16_t completion_space_get(const struct shell *shell)
+{
+	u16_t space = (CONFIG_SHELL_CMD_BUFF_SIZE - 1) -
+			shell->ctx->cmd_buff_len;
+	return space;
+}
+
+/* Prepare arguments and return number of space available for completion. */
+static bool shell_tab_prepare(const struct shell *shell,
+			      const struct shell_static_entry **cmd,
+			      char **argv, size_t *argc,
+			      size_t *complete_arg_idx,
+			      struct shell_static_entry *d_entry)
+{
+	u16_t compl_space = completion_space_get(shell);
+	size_t search_argc;
+
+	if (compl_space == 0) {
+		return false;
+	}
+
+	/* If the Tab key is pressed, "history mode" must be terminated because
+	 * tab and history handlers are sharing the same array: temp_buff.
+	 */
+	history_mode_exit(shell);
+
+	/* Copy command from its beginning to cursor position. */
+	memcpy(shell->ctx->temp_buff, shell->ctx->cmd_buff,
+			shell->ctx->cmd_buff_pos);
+	shell->ctx->temp_buff[shell->ctx->cmd_buff_pos] = '\0';
+
+	/* Create argument list. */
+	(void)shell_make_argv(argc, argv, shell->ctx->temp_buff,
+			      CONFIG_SHELL_ARGC_MAX);
+
+	/* If last command is not completed (followed by space) it is treated
+	 * as uncompleted one.
+	 */
+	int space = isspace((int)shell->ctx->cmd_buff[
+						shell->ctx->cmd_buff_pos - 1]);
+
+	/* root command completion */
+	if ((*argc == 0) || ((space == 0) && (*argc == 1))) {
+		*complete_arg_idx = SHELL_CMD_ROOT_LVL;
+		*cmd = NULL;
+		return true;
+	}
+
+	search_argc = space ? *argc : *argc - 1;;
+
+	*cmd = get_last_command(shell, search_argc, argv, complete_arg_idx,
+				d_entry);
+
+	/* if search_argc == 0 (empty command line) get_last_command will return
+	 * NULL tab is allowed, otherwise not.
+	 */
+	if ((*cmd == NULL) && (search_argc != 0)) {
+		return false;
+	}
+
+	return true;
+}
+
+static bool is_completion_candidate(const char *candidate,
+				    const char *str, size_t len)
+{
+	return (strncmp(candidate, str, len) == 0) ? true : false;
+}
+
+static void find_completion_candidates(const struct shell_static_entry *cmd,
+				       const char *incompl_cmd,
+				       size_t *first_idx, size_t *cnt,
+				       u16_t *longest)
+{
+	struct shell_static_entry dynamic_entry;
+	const struct shell_static_entry *candidate;
+	size_t idx = 0;
+	bool found = false;
+	size_t incompl_cmd_len = shell_strlen(incompl_cmd);
+
+	*longest = 0;
+	*cnt = 0;
+
+	while (true) {
+		cmd_get(cmd ? cmd->subcmd : NULL, cmd ? 1 : 0,
+			idx, &candidate, &dynamic_entry);
+
+		if (!candidate) {
 			break;
 		}
 
-		argv[argc++] = str;
+		if (is_completion_candidate(candidate->syntax, incompl_cmd,
+				incompl_cmd_len)) {
+			size_t slen = strlen(candidate->syntax);
 
-		if (argc == size) {
-			printk("Too many parameters (max %zu)\n", size - 1);
-			return 0;
-		}
-	}
+			*longest = (slen > *longest) ? slen : *longest;
+			(*cnt)++;
 
-	/* keep it POSIX style where argv[argc] is required to be NULL */
-	argv[argc] = NULL;
-
-	return argc;
-}
-
-static struct shell_module *get_destination_module(const char *module_str)
-{
-	int i;
-
-	for (i = 0; i < NUM_OF_SHELL_ENTITIES; i++) {
-		if (!strncmp(module_str,
-			     __shell_module_start[i].module_name,
-			     MODULE_NAME_MAX_LEN)) {
-			return &__shell_module_start[i];
-		}
-	}
-
-	return NULL;
-}
-
-static int show_cmd_help(const struct shell_cmd *cmd, bool full)
-{
-	printk("Usage: %s %s\n", cmd->cmd_name, cmd->help ? cmd->help : "");
-
-	if (full && cmd->desc) {
-		printk("%s\n", cmd->desc);
-	}
-
-	return 0;
-}
-
-static void print_module_commands(struct shell_module *module)
-{
-	int i;
-
-	printk("help\n");
-
-	for (i = 0; module->commands[i].cmd_name; i++) {
-		printk("%-28s %s\n",
-		       module->commands[i].cmd_name,
-		       module->commands[i].help ?
-		       module->commands[i].help : "");
-	}
-}
-
-static const struct shell_cmd *get_cmd(const struct shell_cmd cmds[],
-				       const char *cmd_str)
-{
-	int i;
-
-	for (i = 0; cmds[i].cmd_name; i++) {
-		if (!strcmp(cmd_str, cmds[i].cmd_name)) {
-			return &cmds[i];
-		}
-	}
-
-	return NULL;
-}
-
-static const struct shell_cmd *get_module_cmd(struct shell_module *module,
-					   const char *cmd_str)
-{
-	return get_cmd(module->commands, cmd_str);
-}
-
-static const struct shell_cmd *get_standalone(const char *command)
-{
-	int i;
-
-	for (i = 0; i < NUM_OF_SHELL_CMDS; i++) {
-		if (!strcmp(command, __shell_cmd_start[i].cmd_name)) {
-			return &__shell_cmd_start[i];
-		}
-	}
-
-	return NULL;
-}
-
-/**
- * Handle internal 'help' command
- */
-static int cmd_help(int argc, char *argv[])
-{
-	struct shell_module *module = default_module;
-
-	/* help per command */
-	if (argc > 1) {
-		const struct shell_cmd *cmd;
-		const char *cmd_str;
-
-		module = get_destination_module(argv[1]);
-		if (module) {
-			if (argc == 2) {
-				goto module_help;
+			if (!found) {
+				*first_idx = idx;
 			}
 
-			cmd_str = argv[2];
+			found = true;
 		} else {
-			cmd_str = argv[1];
-			module = default_module;
-		}
-
-		if (!module) {
-			cmd = get_standalone(cmd_str);
-			if (cmd) {
-				return show_cmd_help(cmd, true);
-			} else {
-				printk("No help found for '%s'\n", cmd_str);
-				return -EINVAL;
-			}
-		} else {
-			cmd = get_module_cmd(module, cmd_str);
-			if (cmd) {
-				return show_cmd_help(cmd, true);
-			} else {
-				printk("Unknown command '%s'\n", cmd_str);
-				return -EINVAL;
-			}
-		}
-	}
-
-module_help:
-	/* help per module */
-	if (module) {
-		print_module_commands(module);
-		printk("\nEnter 'exit' to leave current module.\n");
-	} else { /* help for all entities */
-		int i;
-
-		printk("[Modules]\n");
-
-		if (NUM_OF_SHELL_ENTITIES == 0) {
-			printk("No registered modules.\n");
-		}
-
-		for (i = 0; i < NUM_OF_SHELL_ENTITIES; i++) {
-			printk("%s\n", __shell_module_start[i].module_name);
-		}
-
-		printk("\n[Commands]\n");
-
-		if (NUM_OF_SHELL_CMDS == 0) {
-			printk("No registered commands.\n");
-		}
-
-		for (i = 0; i < NUM_OF_SHELL_CMDS; i++) {
-			printk("%s\n", __shell_cmd_start[i].cmd_name);
-		}
-
-		printk("\nTo select a module, enter 'select <module name>'.\n");
-	}
-
-	return 0;
-}
-
-static int set_default_module(const char *name)
-{
-	struct shell_module *module;
-
-	if (strlen(name) > MODULE_NAME_MAX_LEN) {
-		printk("Module name %s is too long, default is not changed\n",
-		       name);
-		return -EINVAL;
-	}
-
-	module = get_destination_module(name);
-
-	if (!module) {
-		printk("Illegal module %s, default is not changed\n", name);
-		return -EINVAL;
-	}
-
-	default_module = module;
-
-	strncpy(default_module_prompt, name, MODULE_NAME_MAX_LEN);
-	strcat(default_module_prompt, "> ");
-
-	return 0;
-}
-
-static int cmd_select(int argc, char *argv[])
-{
-	if (argc == 1) {
-		default_module = NULL;
-		return 0;
-	}
-
-	return set_default_module(argv[1]);
-}
-
-static int cmd_exit(int argc, char *argv[])
-{
-	if (argc == 1) {
-		default_module = NULL;
-	}
-
-	return 0;
-}
-
-static int cmd_noprompt(int argc, char *argv[])
-{
-	no_promt = true;
-	return 0;
-}
-
-#define SHELL_CMD_NOPROMPT "noprompt"
-SHELL_REGISTER_COMMAND(SHELL_CMD_NOPROMPT, cmd_noprompt,
-		       "Disable shell prompt");
-
-static const struct shell_cmd *get_internal(const char *command)
-{
-	static const struct shell_cmd internal_commands[] = {
-		{ "help", cmd_help, "[command]" },
-		{ "select", cmd_select, "[module]" },
-		{ "exit", cmd_exit, NULL },
-		{ NULL },
-	};
-
-	return get_cmd(internal_commands, command);
-}
-
-void shell_register_mcumgr_handler(shell_mcumgr_function_t handler, void *arg)
-{
-	mcumgr_cmd_handler = handler;
-	mcumgr_arg = arg;
-}
-
-int shell_exec(char *line)
-{
-	char *argv[ARGC_MAX + 1], **argv_start = argv;
-	const struct shell_cmd *cmd;
-	int argc, err;
-
-	argc = line2argv(line, argv, ARRAY_SIZE(argv));
-	if (!argc) {
-		return -EINVAL;
-	}
-
-	cmd = get_internal(argv[0]);
-	if (cmd) {
-		goto done;
-	}
-
-	cmd = get_standalone(argv[0]);
-	if (cmd) {
-		goto done;
-	}
-
-	if (argc == 1 && !default_module && NUM_OF_SHELL_CMDS == 0) {
-		printk("No module selected. Use 'select' or 'help'.\n");
-		return -EINVAL;
-	}
-
-	if (default_module) {
-		cmd = get_module_cmd(default_module, argv[0]);
-	}
-
-	if (!cmd && argc > 1) {
-		struct shell_module *module;
-
-		module = get_destination_module(argv[0]);
-		if (module) {
-			cmd = get_module_cmd(module, argv[1]);
-			if (cmd) {
-				argc--;
-				argv_start++;
-			}
-		}
-	}
-
-	if (!cmd) {
-		if (app_cmd_handler) {
-			return app_cmd_handler(argc, argv);
-		}
-
-		printk("Unrecognized command: %s\n", argv[0]);
-		printk("Type 'help' for list of available commands\n");
-		return -EINVAL;
-	}
-
-done:
-	err = cmd->cb(argc, argv_start);
-	if (err < 0) {
-		show_cmd_help(cmd, false);
-	}
-
-	return err;
-}
-
-static void shell(void *p1, void *p2, void *p3)
-{
-	bool skip_prompt = false;
-
-	ARG_UNUSED(p1);
-	ARG_UNUSED(p2);
-	ARG_UNUSED(p3);
-
-	printk("Zephyr Shell, Zephyr version: %s\n", KERNEL_VERSION_STRING);
-	printk("Type 'help' for a list of available commands\n");
-	while (1) {
-		struct console_input *cmd;
-
-		if (!no_promt && !skip_prompt) {
-			printk("%s", get_prompt());
-#if defined(CONFIG_NATIVE_POSIX_CONSOLE)
-			/* The native printk driver is line buffered */
-			posix_flush_stdout();
-#endif
-		}
-
-		cmd = k_fifo_get(&cmds_queue, K_FOREVER);
-
-		/* If the received line is an mcumgr frame, divert it to the
-		 * mcumgr handler.  Don't print the shell prompt this time, as
-		 * that will interfere with the mcumgr response.
-		 */
-		if (mcumgr_cmd_handler != NULL && cmd->is_mcumgr) {
-			mcumgr_cmd_handler(cmd->line, mcumgr_arg);
-			skip_prompt = true;
-		} else {
-			shell_exec(cmd->line);
-			skip_prompt = false;
-		}
-
-		k_fifo_put(&avail_queue, cmd);
-	}
-}
-
-static struct shell_module *get_completion_module(char *str,
-						  char **command_prefix)
-{
-	char dest_str[MODULE_NAME_MAX_LEN];
-	struct shell_module *dest;
-	char *start;
-
-	/* remove ' ' at the beginning of the line */
-	while (*str && *str == ' ') {
-		str++;
-	}
-
-	if (!*str) {
-		return NULL;
-	}
-
-	start = str;
-
-	if (default_module) {
-		dest = default_module;
-		/* caller function already checks str len and put '\0' */
-		*command_prefix = str;
-	} else {
-		dest = NULL;
-	}
-
-	/*
-	 * In case of a default module: only one parameter is possible.
-	 * Otherwise, only two parameters are possibles.
-	 */
-	str = strchr(str, ' ');
-	if (default_module) {
-		return str ? NULL : dest;
-	}
-
-	if (!str) {
-		return NULL;
-	}
-
-	if ((str - start + 1) >= MODULE_NAME_MAX_LEN) {
-		return NULL;
-	}
-
-	strncpy(dest_str, start, (str - start + 1));
-	dest_str[str - start] = '\0';
-	dest = get_destination_module(dest_str);
-	if (!dest) {
-		return NULL;
-	}
-
-	str++;
-
-	/* caller func has already checked str len and put '\0' at the end */
-	*command_prefix = str;
-	str = strchr(str, ' ');
-
-	/* only two parameters are possibles in case of no default module */
-	return str ? NULL : dest;
-}
-
-static u8_t completion(char *line, u8_t len)
-{
-	const char *first_match = NULL;
-	int common_chars = -1, space = 0;
-	int i, command_len;
-	const struct shell_module *module;
-	char *command_prefix;
-
-	if (len >= (MODULE_NAME_MAX_LEN + COMMAND_MAX_LEN - 1)) {
-		return 0;
-	}
-
-	/*
-	 * line to completion is not ended by '\0' as the line that gets from
-	 * k_fifo_get function
-	 */
-	line[len] = '\0';
-	module = get_completion_module(line, &command_prefix);
-	if (!module) {
-		return 0;
-	}
-
-	command_len = strlen(command_prefix);
-
-	for (i = 0; module->commands[i].cmd_name; i++) {
-		int j;
-
-		if (strncmp(command_prefix,
-			    module->commands[i].cmd_name, command_len)) {
-			continue;
-		}
-
-		if (!first_match) {
-			first_match = module->commands[i].cmd_name;
-			continue;
-		}
-
-		/* more commands match, print first match */
-		if (first_match && (common_chars < 0)) {
-			printk("\n%s\n", first_match);
-			common_chars = strlen(first_match);
-		}
-
-		/* cut common part of matching names */
-		for (j = 0; j < common_chars; j++) {
-			if (first_match[j] != module->commands[i].cmd_name[j]) {
+			if (found) {
 				break;
 			}
 		}
+		idx++;
+	}
+}
 
-		common_chars = j;
+static void autocomplete(const struct shell *shell,
+			 const struct shell_static_entry *cmd,
+			 const char *arg,
+			 size_t subcmd_idx)
+{
+	struct shell_static_entry dynamic_entry;
+	const struct shell_static_entry *match;
+	size_t arg_len = shell_strlen(arg);
+	size_t cmd_len;
 
-		printk("%s\n", module->commands[i].cmd_name);
+	cmd_get(cmd ? cmd->subcmd : NULL, cmd ? 1 : 0,
+			subcmd_idx, &match, &dynamic_entry);
+	cmd_len = shell_strlen(match->syntax);
+
+	/* no exact match found */
+	if (cmd_len != arg_len) {
+		shell_op_completion_insert(shell,
+				match->syntax + arg_len,
+				cmd_len - arg_len);
 	}
 
-	/* no match, do nothing */
-	if (!first_match) {
-		return 0;
-	}
-
-	if (common_chars >= 0) {
-		/* multiple match, restore prompt */
-		printk("%s", get_prompt());
-		printk("%s", line);
+	/* Next character in the buffer is not 'space'. */
+	if (!isspace((int) shell->ctx->cmd_buff[
+					shell->ctx->cmd_buff_pos])) {
+		if (shell->ctx->internal.flags.insert_mode) {
+			shell->ctx->internal.flags.insert_mode = 0;
+			shell_op_char_insert(shell, ' ');
+			shell->ctx->internal.flags.insert_mode = 1;
+		} else {
+			shell_op_char_insert(shell, ' ');
+		}
 	} else {
-		common_chars = strlen(first_match);
-		space = 1;
+		/*  case:
+		 | | -> cursor
+		 cons_name $: valid_cmd valid_sub_cmd| |argument  <tab>
+		 */
+		shell_op_cursor_move(shell, 1);
+		/* result:
+		 cons_name $: valid_cmd valid_sub_cmd |a|rgument
+		 */
 	}
-
-	/* complete common part */
-	for (i = command_len; i < common_chars; i++) {
-		printk("%c", first_match[i]);
-		line[len++] = first_match[i];
-	}
-
-	/* for convenience add space after command */
-	if (space) {
-		printk(" ");
-		line[len] = ' ';
-	}
-
-	return common_chars - command_len + space;
 }
 
-
-void shell_init(const char *str)
+static size_t shell_str_common(const char *s1, const char *s2, size_t n)
 {
-	k_fifo_init(&cmds_queue);
-	k_fifo_init(&avail_queue);
+	size_t common = 0;
+	while ((n > 0) && (*s1 == *s2) && (*s1 != '\0')) {
+		s1++;
+		s2++;
+		n--;
+		common++;
+	}
 
-	line_queue_init();
-
-	prompt = str ? str : "";
-
-	k_thread_create(&shell_thread, stack, STACKSIZE, shell, NULL, NULL,
-			NULL, K_PRIO_COOP(7), 0, K_NO_WAIT);
-
-	/* Register console handler */
-	console_register_line_input(&avail_queue, &cmds_queue, completion);
+	return common;
 }
 
-/** @brief Optionally register an app default cmd handler.
- *
- *  @param handler To be called if no cmd found in cmds registered with
- *  shell_init.
+static void options_print(const struct shell *shell,
+			  const struct shell_static_entry *cmd,
+			  size_t first, size_t cnt, u16_t longest)
+{
+	struct shell_static_entry dynamic_entry;
+	const struct shell_static_entry *match;
+	size_t idx = first;
+
+	/* Printing all matching commands (options). */
+	option_print(shell, SHELL_INIT_OPTION_PRINTER, longest);
+
+	while (cnt) {
+		cmd_get(cmd ? cmd->subcmd : NULL, cmd ? 1 : 0,
+			idx, &match, &dynamic_entry);
+		option_print(shell, match->syntax, longest);
+		cnt--;
+		idx++;
+	}
+
+	shell_fprintf(shell, SHELL_INFO, "\r\n%s", shell->name);
+	shell_fprintf(shell, SHELL_NORMAL, "%s", shell->ctx->cmd_buff);
+
+	shell_op_cursor_position_synchronize(shell);
+}
+
+static u16_t common_begining_find(const struct shell_static_entry *cmd,
+				  const char **str,
+				  size_t first, size_t cnt)
+{
+	struct shell_static_entry dynamic_entry;
+	const struct shell_static_entry *match;
+	u16_t common = UINT16_MAX;
+
+
+	cmd_get(cmd ? cmd->subcmd : NULL, cmd ? 1 : 0,
+			first, &match, &dynamic_entry);
+	*str = match->syntax;
+
+	for (size_t idx = first + 1; idx < first + cnt; idx++) {
+		struct shell_static_entry dynamic_entry2;
+		const struct shell_static_entry *match2;
+		int curr_common;
+
+		cmd_get(cmd ? cmd->subcmd : NULL, cmd ? 1 : 0,
+				idx, &match2, &dynamic_entry2);
+
+		curr_common = shell_str_common(match->syntax, match2->syntax,
+					       UINT16_MAX);
+		common = (curr_common < common) ? curr_common : common;
+	}
+
+	return common;
+}
+
+static void partial_autocomplete(const struct shell *shell,
+				    const struct shell_static_entry *cmd,
+				    const char *arg,
+				    size_t first, size_t cnt)
+{
+	const char *completion;
+	u16_t common = common_begining_find(cmd, &completion, first, cnt);
+	int arg_len = shell_strlen(arg);
+
+	if (common) {
+		shell_op_completion_insert(shell, &completion[arg_len],
+					   common - arg_len);
+	}
+}
+
+static void shell_tab_handle(const struct shell *shell)
+{
+	/* +1 reserved for NULL in function shell_make_argv */
+	char *argv[CONFIG_SHELL_ARGC_MAX + 1];
+	/* d_entry - placeholder for dynamic command */
+	struct shell_static_entry d_entry;
+	const struct shell_static_entry *cmd;
+	size_t arg_idx;
+	u16_t longest;
+	size_t first;
+	size_t argc;
+	size_t cnt;
+
+
+	bool tab_possible = shell_tab_prepare(shell, &cmd, argv, &argc,
+					      &arg_idx, &d_entry);
+
+	if (tab_possible == false) {
+		return;
+	}
+
+	find_completion_candidates(cmd, argv[arg_idx], &first, &cnt, &longest);
+
+	if (!cnt) {
+		/* No candidates to propose. */
+		return;
+	} else if (cnt == 1) {
+		/* Autocompletion.*/
+		autocomplete(shell, cmd, argv[arg_idx], first);
+	} else {
+		options_print(shell, cmd, first, cnt, longest);
+		partial_autocomplete(shell, cmd, argv[arg_idx], first, cnt);
+	}
+}
+
+#define SHELL_ASCII_MAX_CHAR (127u)
+static inline int ascii_filter(char const data) {
+	return (u8_t) data > SHELL_ASCII_MAX_CHAR ?
+			-EINVAL : 0;
+}
+
+static void metakeys_handle(const struct shell *shell, char data)
+{
+	/* Optional feature */
+	if (!IS_ENABLED(CONFIG_SHELL_METAKEYS)) {
+		return;
+	}
+
+	switch(data) {
+	case SHELL_VT100_ASCII_CTRL_A: /* CTRL + A */
+		shell_op_cursor_home_move(shell);
+		break;
+
+	case SHELL_VT100_ASCII_CTRL_C: /* CTRL + C */
+		shell_op_cursor_end_move(shell);
+		shell_op_cond_next_line(shell);
+		shell_state_set(shell, SHELL_STATE_ACTIVE);
+		break;
+
+	case SHELL_VT100_ASCII_CTRL_E: /* CTRL + E */
+		shell_op_cursor_end_move(shell);
+		break;
+
+	case SHELL_VT100_ASCII_CTRL_L: /* CTRL + L */
+		SHELL_VT100_CMD(shell, SHELL_VT100_CURSORHOME);
+		SHELL_VT100_CMD(shell, SHELL_VT100_CLEARSCREEN);
+		shell_fprintf(shell, SHELL_INFO, "%s", shell->name);
+		if (flag_echo_is_set(shell))
+		{
+			shell_fprintf(shell, SHELL_NORMAL, "%s",
+				      shell->ctx->cmd_buff);
+			shell_op_cursor_position_synchronize(shell);
+		}
+		break;
+
+	case SHELL_VT100_ASCII_CTRL_U: /* CTRL + U */
+		shell_op_cursor_home_move(shell);
+		cli_cmd_buffer_clear(shell);
+		clear_eos(shell);
+		break;
+
+	case SHELL_VT100_ASCII_CTRL_W: /* CTRL + W */
+		shell_op_word_remove(shell);
+		break;
+	default:
+		break;
+	}
+}
+
+static void cli_state_collect(const struct shell *shell) {
+	size_t count = 0;
+	char data;
+
+	while (1) {
+		(void)shell->iface->api->read(shell->iface, &data,
+					      sizeof(data), &count);
+		if (count == 0) {
+			return;
+		}
+
+		if (ascii_filter(data) != 0) {
+			continue;
+		}
+
+		/* todo pwr_mgmt_feed();*/
+
+		switch (shell->ctx->receive_state) {
+		case SHELL_RECEIVE_DEFAULT:
+			if (data == shell->newline_char) {
+				if (!shell->ctx->cmd_buff_len) {
+					history_mode_exit(shell);
+					cursor_next_line_move(shell);
+				} else {
+					/* Command execution */
+					shell_execute(shell);
+				}
+				shell_state_set(shell, SHELL_STATE_ACTIVE);
+				return;
+			}
+			switch (data) {
+			case SHELL_VT100_ASCII_ESC: /* ESCAPE */
+				receive_state_change(shell, SHELL_RECEIVE_ESC);
+				break;
+
+			case '\0':
+				break;
+
+			case '\t': /* TAB */
+				if (flag_echo_is_set(shell)) {
+					shell_tab_handle(shell);
+				}
+				break;
+
+			case SHELL_VT100_ASCII_BSPACE: /* BACKSPACE */
+				if (flag_echo_is_set(shell)) {
+					shell_op_char_backspace(shell);
+				}
+				break;
+
+			case SHELL_VT100_ASCII_DEL: /* DELETE */
+				if (flag_echo_is_set(shell)) {
+					shell_op_char_delete(shell);
+				}
+				break;
+
+			default:
+				if (isprint((int) data)) {
+					shell_op_char_insert(shell, data);
+				} else {
+					metakeys_handle(shell, data);
+				}
+				break;
+			}
+			break;
+		case SHELL_RECEIVE_ESC:
+			if (data == '[') {
+				receive_state_change(shell,
+						SHELL_RECEIVE_ESC_SEQ);
+			} else {
+				receive_state_change(shell,
+						SHELL_RECEIVE_DEFAULT);
+			}
+			break;
+		case SHELL_RECEIVE_ESC_SEQ:
+			receive_state_change(shell, SHELL_RECEIVE_DEFAULT);
+
+			if (!flag_echo_is_set(shell)) {
+				return;
+			}
+
+			switch (data) {
+			case 'A': /* UP arrow */
+				history_handle(shell, true);
+				break;
+
+			case 'B': /* DOWN arrow */
+				history_handle(shell, false);
+				break;
+
+			case 'C': /* RIGHT arrow */
+				shell_op_right_arrow(shell);
+				break;
+			case 'D': /* LEFT arrow */
+				shell_op_left_arrow(shell);
+				break;
+			case '4': /* END Button in ESC[n~ mode */
+				receive_state_change(shell,
+						SHELL_RECEIVE_TILDE_EXP);
+				/* fall through */
+				/* no break */
+			case 'F': /* END Button in VT100 mode */
+				shell_op_cursor_end_move(shell);
+				break;
+			case '1': /* HOME Button in ESC[n~ mode */
+				receive_state_change(shell,
+						SHELL_RECEIVE_TILDE_EXP);
+				/* fall through */
+				/* no break */
+			case 'H': /* HOME Button in VT100 mode */
+				shell_op_cursor_home_move(shell);
+				break;
+			case '2': /* INSERT Button in ESC[n~ mode */
+				receive_state_change(shell,
+						SHELL_RECEIVE_TILDE_EXP);
+				/* fall through */
+				/* no break */
+			case 'L': /* INSERT Button in VT100 mode */
+				shell->ctx->internal.flags.insert_mode ^= 1;
+				break;
+			case '3':/* DELETE Button in ESC[n~ mode */
+				receive_state_change(shell,
+						SHELL_RECEIVE_TILDE_EXP);
+				if (flag_echo_is_set(shell)) {
+					shell_op_char_delete(shell);
+				}
+				break;
+			default:
+				break;
+			}
+			break;
+		case SHELL_RECEIVE_TILDE_EXP:
+			receive_state_change(shell, SHELL_RECEIVE_DEFAULT);
+			break;
+		default:
+			receive_state_change(shell, SHELL_RECEIVE_DEFAULT);
+			break;
+		}
+	}
+}
+
+static void cmd_trim(const struct shell *shell)
+{
+	shell_buffer_trim(shell->ctx->cmd_buff, &shell->ctx->cmd_buff_len);
+	shell->ctx->cmd_buff_pos = shell->ctx->cmd_buff_len;
+}
+
+/* Function returning pointer to root command matching requested syntax. */
+struct shell_cmd_entry const * root_cmd_find(const char *syntax)
+{
+	const size_t cmd_count = shell_root_cmd_count();
+	struct shell_cmd_entry const *cmd;
+
+	for (size_t cmd_idx = 0; cmd_idx < cmd_count; ++cmd_idx) {
+		cmd = shell_root_cmd_get(cmd_idx);
+		if (strcmp(syntax, cmd->u.entry->syntax) == 0) {
+			return cmd;
+		}
+	}
+
+	return NULL;
+}
+
+/* Function is analyzing the command buffer to find matching commands. Next, it
+ * invokes the  last recognized command which has a handler and passes the rest
+ * of command buffer as arguments.
  */
-void shell_register_app_cmd_handler(shell_cmd_function_t handler)
+static void shell_execute(const struct shell *shell)
 {
-	app_cmd_handler = handler;
+	char quote;
+	size_t argc;
+	/* +1 reserved for NULL added by function shell_make_argv */
+	char *argv[CONFIG_SHELL_ARGC_MAX + 1];
+
+	size_t cmd_idx; /* currently analyzed command in cmd_level */
+	/* currently analyzed command level */
+	size_t cmd_lvl = SHELL_CMD_ROOT_LVL;
+
+	bool wildcard_found = false;
+
+	struct shell_cmd_entry const *p_cmd = NULL;
+
+	struct shell_static_entry const *p_static_entry = NULL;
+
+	struct shell_static_entry d_entry; /* Memory for dynamic commands. */
+	size_t cmd_with_handler_lvl = 0;
+
+
+	memset(&shell->ctx->active_cmd, 0, sizeof(shell->ctx->active_cmd));
+
+
+	cmd_trim(shell);
+
+	history_put(shell, shell->ctx->cmd_buff,
+		    shell->ctx->cmd_buff_len);
+
+	if (IS_ENABLED(CONFIG_SHELL_WILDCARD)) {
+		shell_wildcard_prepare(shell);
+	}
+
+	shell_op_cursor_end_move(shell);
+	cursor_next_line_move(shell);
+
+	/* create argument list */
+	quote = shell_make_argv(&argc, &argv[0], shell->ctx->cmd_buff,
+				CONFIG_SHELL_ARGC_MAX);
+
+	if (!argc) {
+		cursor_next_line_move(shell);
+		return;
+	}
+
+	if (quote != 0) {
+		shell_fprintf(shell, SHELL_ERROR, "not terminated: %c\r\n",
+			      quote);
+		return;
+	}
+
+	/*  Searching for a matching root command. */
+	p_cmd = root_cmd_find(argv[0]);
+	if (p_cmd == NULL) {
+		shell_fprintf(shell, SHELL_ERROR, "%s%s\r\n", argv[0],
+					SHELL_MSG_COMMAND_NOT_FOUND);
+		return;
+	}
+
+	/* Root command shall be always static. */
+	assert(p_cmd->is_dynamic == false);
+
+	/* checking if root command has a handler */
+	if (p_cmd->u.entry->handler != NULL) {
+		shell->ctx->active_cmd = *p_cmd->u.entry;
+	}
+
+	p_cmd = p_cmd->u.entry->subcmd;
+	cmd_lvl++;
+	cmd_idx = 0;
+
+	while (true) {
+		if (cmd_lvl >= argc) {
+			break;
+		}
+
+		if (!strcmp(argv[cmd_lvl], "-h") ||
+		    !strcmp(argv[cmd_lvl], "--help")) {
+			/* Command called with help option so it makes no sense
+			 * to search deeper commands.
+			 */
+			flag_help_set(shell);
+			break;
+		}
+
+		if (IS_ENABLED(CONFIG_SHELL_WILDCARD)) {
+			enum shell_wildcard_status status;
+			status = shell_wildcard_process(shell, p_cmd,
+							 argv[cmd_lvl]);
+			/* Wildcard character found but there is no matching
+			 * command.
+			 */
+			if (status == SHELL_WILDCARD_CMD_NO_MATCH_FOUND) {
+				break;
+			}
+
+			/* Wildcard character was not found function can process
+			 * argument.
+			 */
+			if (status != SHELL_WILDCARD_NOT_FOUND) {
+				++cmd_lvl;
+				wildcard_found = true;
+				continue;
+			}
+		}
+
+		cmd_get(p_cmd, cmd_lvl, cmd_idx++, &p_static_entry, &d_entry);
+
+		if ((cmd_idx == 0) || (p_static_entry == NULL)) {
+			break;
+		}
+
+		if (strcmp(argv[cmd_lvl], p_static_entry->syntax) == 0) {
+			/* checking if command has a handler */
+			if (p_static_entry->handler != NULL) {
+				if (IS_ENABLED(CONFIG_SHELL_WILDCARD)) {
+					if (wildcard_found) {
+						shell_op_cursor_end_move(shell);
+						shell_op_cond_next_line(shell);
+
+						/* An error occurred, fnmatch
+						 * argument cannot be followed
+						 * by argument with a handler to
+						 * avoid multiple function
+						 * calls.
+						 */
+						shell_fprintf(shell,SHELL_ERROR,
+							"Error: requested "
+							"multiple function "
+							"executions\r\n");
+						flag_help_clear(shell);
+
+						return;
+					}
+				}
+				shell->ctx->active_cmd = *p_static_entry;
+				cmd_with_handler_lvl = cmd_lvl;
+			}
+			cmd_lvl++;
+			cmd_idx = 0;
+			p_cmd = p_static_entry->subcmd;
+		}
+	}
+
+	if (IS_ENABLED(CONFIG_SHELL_WILDCARD)) {
+		shell_wildcard_finalize(shell);
+		/* calling make_arg function again because cmd_buffer was
+		 * overwritten by temporary buffer.
+		 */
+		(void)shell_make_argv(&argc, &argv[0],
+				      shell->ctx->cmd_buff,
+				      CONFIG_SHELL_ARGC_MAX);
+	}
+
+	/* Executing the deepest found handler. */
+	if (shell->ctx->active_cmd.handler == NULL) {
+		shell_fprintf(shell, SHELL_ERROR, SHELL_MSG_SPECIFY_SUBCOMMAND);
+	}
+	else {
+		shell->ctx->active_cmd.handler(shell,
+					       argc - cmd_with_handler_lvl,
+					       &argv[cmd_with_handler_lvl]);
+	}
+
+	flag_help_clear(shell);
 }
 
-void shell_register_prompt_handler(shell_prompt_function_t handler)
+static void shell_transport_evt_handler(enum shell_transport_evt evt_type,
+				      void *context)
 {
-	app_prompt_handler = handler;
+	struct shell *shell = (struct shell *)context;
+	struct k_poll_signal *signal;
+
+	signal = (evt_type == SHELL_TRANSPORT_EVT_RX_RDY) ?
+			&shell->ctx->signals[SHELL_SIGNAL_RXRDY] :
+			&shell->ctx->signals[SHELL_SIGNAL_TXDONE];
+	k_poll_signal(signal, 0);
 }
 
-void shell_register_default_module(const char *name)
+static void shell_current_command_erase(const struct shell *shell)
 {
-	int err = set_default_module(name);
+	shell_multiline_data_calc(&shell->ctx->vt100_ctx.cons,
+				  shell->ctx->cmd_buff_pos,
+				  shell->ctx->cmd_buff_len);
+	//shell_op_cursor_horiz_move(shell, -shell->ctx->vt100_ctx.cons.cur_y);
+	shell_op_cursor_horiz_move(shell, -shell->ctx->vt100_ctx.cons.cur_x);
+	clear_eos(shell);
+}
 
-	if (!err) {
-		printk("\n%s", default_module_prompt);
+static void shell_current_command_print(const struct shell *shell)
+{
+	shell_fprintf(shell, SHELL_INFO, "%s", shell->name);
+
+	if (flag_echo_is_set(shell)) {
+		shell_fprintf(shell, SHELL_NORMAL, "%s", shell->ctx->cmd_buff);
+		shell_op_cursor_position_synchronize(shell);
 	}
 }
+
+static void shell_log_process(const struct shell *shell)
+{
+	bool processed;
+	int signaled;
+	int result;
+
+	do {
+
+		shell_current_command_erase(shell);
+		processed = shell_log_backend_process(shell->log_backend);
+		shell_current_command_print(shell);
+
+		k_poll_signal_check(&shell->ctx->signals[SHELL_SIGNAL_RXRDY],
+						    &signaled, &result);
+
+	} while (processed && !signaled);
+}
+
+static int shell_instance_init(const struct shell *shell, void const *p_config,
+			       bool use_colors) {
+	assert(shell);
+	assert(shell->ctx && shell->iface && shell->name);
+	assert((shell->newline_char == '\n') || (shell->newline_char == '\r'));
+
+	int err;
+
+	err = shell->iface->api->init(shell->iface, p_config,
+				      shell_transport_evt_handler,
+				      (void *) shell);
+	if (err != 0) {
+		return err;
+	}
+
+	history_init(shell);
+
+	memset(shell->ctx, 0, sizeof(*shell->ctx));
+
+	if (IS_ENABLED(CONFIG_SHELL_STATS)) {
+		shell->stats->log_lost_cnt = 0;
+	}
+
+	shell->ctx->internal.flags.tx_rdy = 1;
+	shell->ctx->internal.flags.echo = CONFIG_SHELL_ECHO_STATUS;
+	shell->ctx->state = SHELL_STATE_INITIALIZED;
+	shell->ctx->vt100_ctx.cons.terminal_wid = SHELL_DEFAULT_TERMINAL_WIDTH;
+	shell->ctx->vt100_ctx.cons.terminal_hei = SHELL_DEFAULT_TERMINAL_HEIGHT;
+	shell->ctx->vt100_ctx.cons.name_len = shell_strlen(shell->name);
+	shell->ctx->internal.flags.use_colors =
+					IS_ENABLED(CONFIG_SHELL_VT100_COLORS);
+
+	return 0;
+}
+
+static int shell_instance_uninit(const struct shell *shell);
+
+void shell_thread(void *shell_handle, void *dummy1, void *dummy2)
+{
+	struct shell *shell = (struct shell *)shell_handle;
+	int err;
+	int i;
+
+	for (i = 0; i < SHELL_SIGNALS; i++) {
+		k_poll_signal_init(&shell->ctx->signals[i]);
+		k_poll_event_init(&shell->ctx->events[i],
+				  K_POLL_TYPE_SIGNAL,
+				  K_POLL_MODE_NOTIFY_ONLY,
+				  &shell->ctx->signals[i]);
+	}
+
+	err = shell_start(shell);
+	if (err != 0) {
+		return;
+	}
+
+	while (1)
+	{
+		int signaled;
+		int result;
+
+		err = k_poll(shell->ctx->events, SHELL_SIGNALS, K_FOREVER);
+		(void)err;
+
+		k_poll_signal_check(&shell->ctx->signals[SHELL_SIGNAL_KILL],
+				    &signaled, &result);
+
+		if (signaled) {
+			k_poll_signal_reset(
+				&shell->ctx->signals[SHELL_SIGNAL_KILL]);
+			(void)shell_instance_uninit(shell);
+
+			k_thread_abort(k_current_get());
+		}
+
+		k_poll_signal_check(&shell->ctx->signals[SHELL_SIGNAL_LOG_MSG],
+				    &signaled, &result);
+
+		if (!signaled) {
+			/* Other signals handled together.*/
+			k_poll_signal_reset(
+				&shell->ctx->signals[SHELL_SIGNAL_RXRDY]);
+			k_poll_signal_reset(
+				&shell->ctx->signals[SHELL_SIGNAL_TXDONE]);
+			shell_process(shell);
+		} else if (IS_ENABLED(CONFIG_LOG)){
+			k_poll_signal_reset(
+				&shell->ctx->signals[SHELL_SIGNAL_LOG_MSG]);
+			/* process log msg */
+			shell_log_process(shell);
+		}
+	}
+}
+
+int shell_init(const struct shell *shell, void const *transport_config,
+	       bool use_colors, bool log_backend, u32_t init_log_level)
+{
+	assert(shell);
+	int err;
+
+	err = shell_instance_init(shell, transport_config, use_colors);
+	if (err != 0) {
+		return err;
+	}
+
+	if (log_backend)
+	{
+		if (IS_ENABLED(CONFIG_LOG)) {
+			shell_log_backend_enable(shell->log_backend,
+						 (void *)shell, init_log_level);
+		}
+	}
+
+	(void)k_thread_create(shell->thread, shell->stack,
+			      CONFIG_SHELL_STACK_SIZE, shell_thread,
+			      (void *)shell, NULL, NULL,
+			      CONFIG_SHELL_THREAD_PRIO, 0, K_NO_WAIT);
+
+	return 0;
+}
+
+static int shell_instance_uninit(const struct shell *shell)
+{
+	assert(shell);
+	assert(shell->ctx && shell->iface && shell->name);
+	int err;
+
+	if (flag_processing_is_set(shell)) {
+		return -EBUSY;
+	}
+
+	if (IS_ENABLED(CONFIG_LOG)) {
+		/* todo purge log queue */
+		shell_log_backend_disable(shell->log_backend);
+	}
+
+	err = shell->iface->api->uninit(shell->iface);
+	if (err != 0) {
+		return err;
+	}
+
+	history_purge(shell);
+
+	shell->ctx->state = SHELL_STATE_UNINITIALIZED;
+
+	return 0;
+}
+
+int shell_uninit(const struct shell *shell)
+{
+	if (IS_ENABLED(CONFIG_MULTITHREADING)) {
+		/* signal kill message */
+		(void)k_poll_signal(&shell->ctx->signals[SHELL_SIGNAL_KILL], 0);
+
+		return 0;
+	} else {
+		return shell_instance_uninit(shell);
+	}
+}
+
+int shell_start(const struct shell *shell)
+{
+	assert(shell);
+	assert(shell->ctx && shell->iface && shell->name);
+	int err;
+
+	if (shell->ctx->state != SHELL_STATE_INITIALIZED) {
+		return -ENOTSUP;
+	}
+
+	err = shell->iface->api->enable(shell->iface, false);
+	if (err != 0) {
+		return err;
+	}
+
+	if (IS_ENABLED(CONFIG_SHELL_VT100_COLORS_ENABLED)) {
+		vt100_color_set(shell, SHELL_NORMAL);
+		vt100_bgcolor_set(shell, SHELL_VT100_COLOR_BLACK);
+	}
+
+	shell_raw_fprintf(shell->fprintf_ctx, "\r\n\n");
+
+	shell_state_set(shell, SHELL_STATE_ACTIVE);
+
+	return 0;
+}
+
+int shell_stop(const struct shell *shell)
+{
+	assert(shell);
+
+	if ((shell->ctx->state == SHELL_STATE_INITIALIZED) ||
+	    (shell->ctx->state == SHELL_STATE_UNINITIALIZED)) {
+		return -ENOTSUP;
+	}
+
+	shell_state_set(shell, SHELL_STATE_INITIALIZED);
+
+	return 0;
+}
+
+void shell_process(const struct shell *shell) {
+	assert(shell);
+
+	union shell_internal internal;
+	internal.value = 0;
+	internal.flags.processing = 1;
+
+	(void)atomic_or((atomic_t *)&shell->ctx->internal.value,
+			internal.value);
+
+	switch (shell->ctx->state) {
+	case SHELL_STATE_UNINITIALIZED:
+	case SHELL_STATE_INITIALIZED:
+		/* Console initialized but not started. */
+		break;
+
+	case SHELL_STATE_ACTIVE:
+		cli_state_collect(shell);
+		break;
+
+	default:
+		break;
+	}
+
+	transport_buffer_flush(shell);
+
+	internal.value = 0xFFFFFFFF;
+	internal.flags.processing = 0;
+	(void)atomic_and((atomic_t *)&shell->ctx->internal.value,
+			 internal.value);
+}
+
+/* Function shall be only used by the fprintf module. */
+void shell_print_stream(void const *user_ctx, char const *data,
+			size_t data_len)
+{
+	shell_write((const struct shell *) user_ctx, data, data_len);
+}
+
+void shell_fprintf(const struct shell *shell, enum shell_vt100_color color,
+		   char const *p_fmt, ...)
+{
+	assert(shell);
+
+	va_list args = { 0 };
+	va_start(args, p_fmt);
+
+	if (IS_ENABLED(CONFIG_SHELL_VT100_COLORS) &&
+	    shell->ctx->internal.flags.use_colors &&
+	    (color != shell->ctx->vt100_ctx.col.col)) {
+		struct shell_vt100_colors col;
+
+		vt100_colors_store(shell, &col);
+		vt100_color_set(shell, color);
+
+		shell_fprintf_fmt(shell->fprintf_ctx, p_fmt, args);
+
+		vt100_colors_restore(shell, &col);
+	} else {
+		shell_fprintf_fmt(shell->fprintf_ctx, p_fmt, args);
+	}
+
+	va_end(args);
+}
+
+/* Function prints a string on terminal screen with requested margin.
+ * It takes care to not divide words.
+ *   shell		Pointer to CLI instance.
+ *   p_str		Pointer to string to be printed.
+ *   terminal_offset	Requested left margin.
+ *   offset_first_line	Add margin to the first printed line.
+ */
+static void format_offset_string_print(const struct shell *shell,
+				       char const *str,
+				       size_t terminal_offset,
+				       bool offset_first_line)
+{
+	size_t length;
+	size_t offset = 0;
+
+	if (str == NULL) {
+		return;
+	}
+
+	if (offset_first_line) {
+		shell_op_cursor_horiz_move(shell, terminal_offset);
+	}
+
+
+	/* Skipping whitespace. */
+	while (isspace((int) *(str + offset))) {
+		++offset;
+	}
+
+	while (1) {
+		size_t idx = 0;
+		length = shell_strlen(str) - offset;
+
+		if (length <=
+		    shell->ctx->vt100_ctx.cons.terminal_wid - terminal_offset) {
+			for (idx = 0; idx < length; idx++) {
+				if (*(str + offset + idx) == '\n') {
+					transport_buffer_flush(shell);
+					shell_write(shell, str + offset, idx);
+					offset += idx + 1;
+					cursor_next_line_move(shell);
+					shell_op_cursor_horiz_move(shell,
+							terminal_offset);
+					break;
+				}
+			}
+
+			/* String will fit in one line. */
+			shell_raw_fprintf(shell->fprintf_ctx, str + offset);
+
+			break;
+		} else {
+			/* String is longer than terminal line so text needs to
+			 * divide in the way to not divide words.
+			 */
+			length = shell->ctx->vt100_ctx.cons.terminal_wid
+					- terminal_offset;
+
+			while (1) {
+				/* Determining line break. */
+				if (isspace((int) (*(str + offset + idx)))) {
+					length = idx;
+					if (*(str + offset + idx) == '\n') {
+						break;
+					}
+				}
+
+				if ((idx + terminal_offset) >=
+				    shell->ctx->vt100_ctx.cons.terminal_wid) {
+					/* End of line reached. */
+					break;
+				}
+
+				++idx;
+			}
+
+			/* Writing one line, fprintf IO buffer must be flushed
+			 * before calling cli_write.
+			 */
+			transport_buffer_flush(shell);
+			shell_write(shell, str + offset, length);
+			offset += length;
+
+			/* Calculating text offset to ensure that next line will
+			 * not begin with a space.
+			 */
+			while (isspace((int) (*(str + offset)))) {
+				++offset;
+			}
+
+			cursor_next_line_move(shell);
+			shell_op_cursor_horiz_move(shell, terminal_offset);
+		}
+	}
+	cursor_next_line_move(shell);
+}
+
+void shell_help_print(const struct shell *shell,
+		      struct shell_getopt_option const *opt, size_t opt_len) {
+	assert(shell);
+
+	if (!IS_ENABLED(CONFIG_SHELL_HELP)) {
+		return;
+	}
+
+	static u8_t const tab_len = 2;
+	static char const opt_sep[] = ", "; /* options separator */
+	static char const help[] = "-h, --help";
+	static char const cmd_sep[] = " - "; /* command separator */
+	u16_t field_width = 0;
+	u16_t longest_string = shell_strlen(help) - shell_strlen(opt_sep);
+
+	/* Printing help string for command. */
+	shell_fprintf(shell, SHELL_NORMAL, "%s%s",
+		      shell->ctx->active_cmd.syntax, cmd_sep);
+
+	field_width = shell_strlen(shell->ctx->active_cmd.syntax)
+			+ shell_strlen(cmd_sep);
+	format_offset_string_print(shell, shell->ctx->active_cmd.help,
+			field_width, false);
+
+	shell_fprintf(shell, SHELL_NORMAL, "Options:\r\n");
+
+	/* Looking for the longest option string. */
+	if ((opt_len > 0) && (opt != NULL)) {
+		for (size_t i = 0; i < opt_len; ++i) {
+			if (shell_strlen(opt[i].optname_short)
+					+ shell_strlen(opt[i].optname)
+					> longest_string) {
+				longest_string =
+					shell_strlen(opt[i].optname_short) +
+						shell_strlen(opt[i].optname);
+			}
+		}
+	}
+
+	longest_string += shell_strlen(opt_sep) + tab_len;
+
+	shell_fprintf(shell,
+	SHELL_NORMAL, "  %-*s:", longest_string, help);
+
+	/* Print help string for options (only -h and --help).
+	 * tab_len + 1 == "  " and ':' from: "  %-*s:"
+	 */
+	field_width = longest_string + tab_len + 1;
+	format_offset_string_print(shell, "Show command help.", field_width,
+				   false);
+
+	/* Formating and printing all available options (except -h, --help). */
+	if (opt) {
+		for (size_t i = 0; i < opt_len; ++i) {
+			if (opt[i].optname_short && opt[i].optname) {
+				shell_fprintf(shell, SHELL_NORMAL,
+					      "  %s%s%s", opt[i].optname_short,
+					      opt_sep, opt[i].optname);
+				field_width = longest_string + tab_len;
+				shell_op_cursor_horiz_move(shell, field_width
+					- (shell_strlen(opt[i].optname_short)
+					   + shell_strlen(opt[i].optname)
+					   + tab_len
+					   + shell_strlen(opt_sep)));
+				shell_putc(shell, ':');
+				/* incrementing because char ':' was already
+				 * printed above
+				 */
+				++field_width;
+			} else if (opt[i].optname_short) {
+				shell_fprintf(shell, SHELL_NORMAL, "  %-*s:",
+					      longest_string,
+					      opt[i].optname_short);
+			/* tab_len + 1 == "  " and ':' from: "  %-*s:" */
+				field_width = longest_string + tab_len + 1;
+			} else if (opt[i].optname) {
+				shell_fprintf(shell, SHELL_NORMAL, "  %-*s:",
+						longest_string, opt[i].optname);
+			/* tab_len + 1 == "  " and ':' from: "  %-*s:" */
+				field_width = longest_string + tab_len + 1;
+			} else {
+				/* Do nothing. */
+			}
+
+			if (opt[i].optname_help) {
+				format_offset_string_print(shell,
+						opt[i].optname_help,
+						field_width, false);
+			} else {
+				cursor_next_line_move(shell);
+			}
+		}
+	}
+
+	/* Checking if there are any subcommands avilable. */
+	if (!shell->ctx->active_cmd.subcmd) {
+		return;
+	}
+
+	/* Printing formatted help of one level deeper subcommands. */
+	struct shell_static_entry static_entry;
+	struct shell_cmd_entry const *cmd = shell->ctx->active_cmd.subcmd;
+	struct shell_static_entry const *st_cmd = NULL;
+	size_t cmd_idx = 0;
+
+	field_width = 0;
+	longest_string = 0;
+
+	/* Searching for the longest subcommand to print. */
+	while (1) {
+		cmd_get(cmd, !SHELL_CMD_ROOT_LVL, cmd_idx++, &st_cmd,
+				&static_entry);
+
+		if (!st_cmd) {
+			break;
+		}
+
+		if (shell_strlen(st_cmd->syntax) > longest_string) {
+			longest_string = shell_strlen(st_cmd->syntax);
+		}
+	}
+
+	/* Checking if there are dynamic subcommands. */
+	if (cmd_idx == 1) {
+		/* No dynamic subcommands available. */
+		return;
+	}
+
+	shell_fprintf(shell, SHELL_NORMAL, "Subcommands:\r\n");
+
+	/* Printing subcommands and help string (if exists). */
+	cmd_idx = 0;
+
+	while (1) {
+		cmd_get(cmd, !SHELL_CMD_ROOT_LVL, cmd_idx++, &st_cmd,
+				&static_entry);
+
+		if (st_cmd == NULL) {
+			break;
+		}
+
+		field_width = longest_string + tab_len;
+		shell_fprintf(shell, SHELL_NORMAL, "  %-*s:", field_width,
+				st_cmd->syntax);
+		/* tab_len + 1 == "  " and ':' from: "  %-*s:" */
+		field_width += tab_len + 1;
+
+		if (st_cmd->help) {
+			format_offset_string_print(shell, st_cmd->help,
+						   field_width, false);
+		} else {
+			cursor_next_line_move(shell);
+		}
+	}
+}
+
+bool shell_cmd_precheck(const struct shell *shell,
+			bool arg_cnt_ok,
+		 	struct shell_getopt_option const *opt,
+		 	size_t opt_len)
+{
+	if (shell_help_requested(shell)) {
+		shell_help_print(shell, opt, opt_len);
+		return true;
+	}
+
+	if (!arg_cnt_ok) {
+		shell_fprintf(shell, SHELL_ERROR,
+			      "%s: wrong parameter count\r\n",
+			      shell->ctx->active_cmd.syntax);
+
+		if (IS_ENABLED(SHELL_HELP_ON_WRONG_ARGUMENT_COUNT)) {
+			shell_help_print(shell, opt, opt_len);
+		}
+
+		return false;
+	}
+
+	return true;
+}
+
+#if 0
+#if SHELL_LOG_BACKEND && NRF_MODULE_ENABLED(NRF_LOG)
+
+#define SHELL_LOG_MSG_OVERFLOW_MSK ((u32_t)7)
+static bool cli_log_entry_process(const struct shell *shell, bool skip) {
+	nrf_log_entry_t entry;
+
+#if NRF_MODULE_ENABLED(SHELL_STATISTICS)
+	bool print_msg = false;
+#endif
+
+	if (nrf_queue_pop(
+			((shell_log_backend_t *) shell->p_log_backend->ctx)->p_queue,
+			&entry) != NRF_SUCCESS) {
+		return false;
+	}
+
+	if (skip) {
+		nrf_memobj_put(entry);
+#if NRF_MODULE_ENABLED(SHELL_STATISTICS)
+		++shell->ctx->statistics.log_lost_cnt;
+		if ((shell->ctx->statistics.log_lost_cnt & SHELL_LOG_MSG_OVERFLOW_MSK) == 1) {
+			/* Set flag to print a message after clearing the currently entered command. */
+			print_msg = true;
+		} else
+#endif
+		{
+			return true;
+		}
+	}
+	{
+		/* Erasing the currently displayed command and console name. */
+		struct shell_multiline_cons *cons = &shell->ctx->vt100_ctx.cons;
+
+		shell_multiline_data_calc(cons, shell->ctx->cmd_buff_pos,
+					  shell->ctx->cmd_buff_len);
+
+		if (cons->cur_y > SHELL_INITIAL_CURS_POS) {
+			cursor_up_move(shell,
+					cons->cur_y - SHELL_INITIAL_CURS_POS);
+		}
+
+		if (cons->cur_x > SHELL_INITIAL_CURS_POS) {
+			cursor_left_move(shell,
+					cons->cur_x - SHELL_INITIAL_CURS_POS);
+		}
+		clear_eos(shell);
+	}
+
+#if NRF_MODULE_ENABLED(SHELL_STATISTICS)
+	if (print_msg) {
+		/* Print the requested string and exit function. */
+		shell_fprintf(shell, SHELL_ERROR, "Lost logs - increase log backend queue size.\r\n");
+
+		return true;
+	}
+#endif
+
+	/* Printing logs from the queue. */
+	do {
+		nrf_log_header_t header;
+		size_t memobj_offset = 0;
+		nrf_log_str_formatter_entry_params_t params;
+
+		nrf_memobj_read(entry, &header, HEADER_SIZE * sizeof(u32_t),
+				memobj_offset);
+		memobj_offset = HEADER_SIZE * sizeof(u32_t);
+
+		params.timestamp = header.timestamp;
+		params.module_id = header.module_id;
+		params.dropped = header.dropped;
+		params.use_colors = NRF_LOG_USES_COLORS; /* Color will be provided by the console application. */
+
+		if (header.base.generic.type == HEADER_TYPE_STD) {
+			char const * p_log_str =
+					(char const *) ((u32_t) header.base.std.addr);
+			params.severity =
+					(nrf_log_severity_t) header.base.std.severity;
+			u32_t nargs = header.base.std.nargs;
+			u32_t args[6];
+
+			nrf_memobj_read(entry, args, nargs * sizeof(u32_t),
+					memobj_offset);
+			nrf_log_std_entry_process(p_log_str, args, nargs,
+					&params, shell->p_fprintf_ctx);
+		} else if (header.base.generic.type == HEADER_TYPE_HEXDUMP) {
+			u32_t data_len;
+			u8_t data_buf[8];
+			u32_t chunk_len;
+
+			data_len = header.base.hexdump.len;
+			params.severity =
+					(nrf_log_severity_t) header.base.hexdump.severity;
+
+			do {
+				chunk_len = sizeof(data_buf) > data_len ?
+						data_len : sizeof(data_buf);
+				nrf_memobj_read(entry, data_buf, chunk_len,
+						memobj_offset);
+				memobj_offset += chunk_len;
+				data_len -= chunk_len;
+				nrf_log_hexdump_entry_process(data_buf,
+						chunk_len, &params,
+						shell->p_fprintf_ctx);
+			} while (data_len > 0);
+		}
+
+		nrf_memobj_put(entry);
+	} while (nrf_queue_pop(
+			((shell_log_backend_t *) shell->p_log_backend->ctx)->p_queue,
+			&entry) == NRF_SUCCESS);
+	return true;
+}
+
+static void nrf_log_backend_cli_put(nrf_log_backend_t const * p_backend,
+		nrf_log_entry_t * p_msg) {
+	shell_log_backend_t * p_backend_cli =
+			(shell_log_backend_t *) p_backend->ctx;
+	shell_t const *shell = p_backend_cli->shell;
+
+	//If panic mode cannot be handled, stop handling new requests.
+	if (shell->ctx->state != SHELL_STATE_PANIC_MODE_INACTIVE) {
+		bool panic_mode = (shell->ctx->state
+				== SHELL_STATE_PANIC_MODE_ACTIVE);
+		//If there is no place for a new log entry, remove the oldest one.
+		ret_code_t err_code = nrf_queue_push(p_backend_cli->p_queue,
+				&p_msg);
+		while (err_code != NRF_SUCCESS) {
+			(void) cli_log_entry_process(shell,
+					panic_mode ? false : true);
+
+			err_code = nrf_queue_push(p_backend_cli->p_queue,
+					&p_msg);
+		}
+		nrf_memobj_get(p_msg);
+
+		if (panic_mode) {
+			(void) cli_log_entry_process(shell, false);
+		}
+#if NRF_MODULE_ENABLED(SHELL_USES_TASK_MANAGER)
+		else {
+			task_events_set((task_id_t)((u32_t)p_backend_cli->p_context & 0x000000FF),
+					SHELL_LOG_PENDING_TASK_EVT);
+		}
+#endif
+	}
+}
+
+static void nrf_log_backend_cli_flush(nrf_log_backend_t const * p_backend) {
+	shell_log_backend_t * p_backend_cli =
+			(shell_log_backend_t *) p_backend->ctx;
+	shell_t const * shell = p_backend_cli->shell;
+	nrf_log_entry_t * p_msg;
+
+	if (nrf_queue_pop(p_backend_cli->p_queue, &p_msg) == NRF_SUCCESS) {
+		(void) cli_log_entry_process(shell, false);
+	}
+	UNUSED_PARAMETER(p_backend);
+}
+
+static void nrf_log_backend_cli_panic_set(nrf_log_backend_t const * p_backend) {
+	shell_log_backend_t * p_backend_cli =
+			(shell_log_backend_t *) p_backend->ctx;
+	shell_t const *shell = p_backend_cli->shell;
+
+	if (shell->p_iface->p_api->enable(shell->p_iface, true)
+			== NRF_SUCCESS) {
+		shell->ctx->state = SHELL_STATE_PANIC_MODE_ACTIVE;
+	} else {
+		shell->ctx->state = SHELL_STATE_PANIC_MODE_INACTIVE;
+	}
+}
+
+const nrf_log_backend_api_t nrf_log_backend_cli_api = { .put =
+		nrf_log_backend_cli_put, .flush = nrf_log_backend_cli_flush,
+		.panic_set = nrf_log_backend_cli_panic_set, };
+#else
+static bool cli_log_entry_process(const struct shell *shell, bool skip)
+{
+	UNUSED_PARAMETER(shell);
+	UNUSED_PARAMETER(skip);
+	return false;
+}
+#endif // SHELL_LOG_BACKEND
+#endif
+const struct log_backend_api log_backend_shell_api;
+
