+++ subsys/shell/shell_wildcard.h
@@ -7,7 +7,7 @@
 #ifndef SHELL_SHELL_WILDCARDS_H__
 #define SHELL_SHELL_WILDCARDS_H__
 
-#include <shell/cli.h>
+#include <shell/shell.h>
 
 enum shell_wildcard_status {
 	SHELL_WILDCARD_CMD_ADDED,
